+++ example/lib/main.dart
@@ -1,69 +1,489 @@
-// FIREBASE_MESSAGING will be discontinued
-// import 'package:firebase_core/firebase_core.dart';
-// import 'package:firebase_messaging/firebase_messaging.dart';
-// import 'package:awesome_notifications_example/notifications/firebase_controller.dart';
-
 import 'package:awesome_notifications/awesome_notifications.dart';
 import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
 
-import 'package:awesome_notifications_example/routes/routes.dart';
-import 'package:awesome_notifications_example/themes/themes_controller.dart';
-import 'package:awesome_notifications_example/notifications/notifications_controller.dart';
-import 'package:flutter/services.dart';
-
-void main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-  NotificationsController.initializeLocalNotifications();
-  App.initialCallAction =
-      await NotificationsController.interceptInitialCallActionRequest();
-
-  await SystemChrome.setPreferredOrientations([
-    DeviceOrientation.portraitUp,
-  ]);
-
-  runApp(const App());
+Future<void> main() async {
+  // Always initialize Awesome Notifications
+  await NotificationController.initializeLocalNotifications();
+  runApp(const MyApp());
 }
 
-class App extends StatefulWidget {
-  const App({super.key});
+///  *********************************************
+///     NOTIFICATION CONTROLLER
+///  *********************************************
+///
+class NotificationController {
+  static ReceivedAction? initialAction;
 
-  static ReceivedAction? initialCallAction;
-  static String name = 'Awesome Notifications - Example App';
-  static Color mainColor = const Color(0xFF9D50DD);
+  ///  *********************************************
+  ///     INITIALIZATIONS
+  ///  *********************************************
+  ///
+  static Future<void> initializeLocalNotifications() async {
+    await AwesomeNotifications().initialize(
+        null, //'resource://drawable/res_app_icon',//
+        [
+          NotificationChannel(
+              channelKey: 'alerts',
+              channelName: 'Alerts',
+              channelDescription: 'Notification tests as alerts',
+              playSound: true,
+              onlyAlertOnce: true,
+              groupAlertBehavior: GroupAlertBehavior.Children,
+              importance: NotificationImportance.High,
+              defaultPrivacy: NotificationPrivacy.Private,
+              defaultColor: Colors.deepPurple,
+              ledColor: Colors.deepPurple)
+        ],
+        debug: true);
+
+    // Get initial notification action is optional
+    initialAction = await AwesomeNotifications()
+        .getInitialNotificationAction(removeFromActionEvents: false);
+  }
+
+  ///  *********************************************
+  ///     NOTIFICATION EVENTS LISTENER
+  ///  *********************************************
+  ///  Notifications events are only delivered after call this method
+  static Future<void> startListeningNotificationEvents() async {
+    AwesomeNotifications()
+        .setListeners(onActionReceivedMethod: onActionReceivedMethod);
+  }
+
+  ///  *********************************************
+  ///     NOTIFICATION EVENTS
+  ///  *********************************************
+  ///
+  @pragma('vm:entry-point')
+  static Future<void> onActionReceivedMethod(
+      ReceivedAction receivedAction) async {
+
+    if(
+      receivedAction.actionType == ActionType.SilentAction ||
+      receivedAction.actionType == ActionType.SilentBackgroundAction
+    ){
+      // For background actions, you must hold the execution until the end
+      print('Message sent via notification input: "${receivedAction.buttonKeyInput}"');
+      await executeLongTaskInBackground();
+    }
+    else {
+      MyApp.navigatorKey.currentState?.pushNamedAndRemoveUntil(
+          '/notification-page',
+              (route) =>
+          (route.settings.name != '/notification-page') || route.isFirst,
+          arguments: receivedAction);
+    }
+  }
+
+  ///  *********************************************
+  ///     REQUESTING NOTIFICATION PERMISSIONS
+  ///  *********************************************
+  ///
+  static Future<bool> displayNotificationRationale() async {
+    bool userAuthorized = false;
+    BuildContext context = MyApp.navigatorKey.currentContext!;
+    await showDialog(
+        context: context,
+        builder: (BuildContext ctx) {
+          return AlertDialog(
+            title: Text('Get Notified!',
+                style: Theme.of(context).textTheme.titleLarge),
+            content: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Row(
+                  children: [
+                    Expanded(
+                      child: Image.asset(
+                        'assets/animated-bell.gif',
+                        height: MediaQuery.of(context).size.height * 0.3,
+                        fit: BoxFit.fitWidth,
+                      ),
+                    ),
+                  ],
+                ),
+                const SizedBox(height: 20),
+                const Text(
+                    'Allow Awesome Notifications to send you beautiful notifications!'),
+              ],
+            ),
+            actions: [
+              TextButton(
+                  onPressed: () {
+                    Navigator.of(ctx).pop();
+                  },
+                  child: Text(
+                    'Deny',
+                    style: Theme.of(context)
+                        .textTheme
+                        .titleLarge
+                        ?.copyWith(color: Colors.red),
+                  )),
+              TextButton(
+                  onPressed: () async {
+                    userAuthorized = true;
+                    Navigator.of(ctx).pop();
+                  },
+                  child: Text(
+                    'Allow',
+                    style: Theme.of(context)
+                        .textTheme
+                        .titleLarge
+                        ?.copyWith(color: Colors.deepPurple),
+                  )),
+            ],
+          );
+        });
+    return userAuthorized &&
+        await AwesomeNotifications().requestPermissionToSendNotifications();
+  }
+
+  ///  *********************************************
+  ///     BACKGROUND TASKS TEST
+  ///  *********************************************
+  static Future<void> executeLongTaskInBackground() async {
+    print("starting long task");
+    await Future.delayed(const Duration(seconds: 4));
+    final url = Uri.parse("http://google.com");
+    final re = await http.get(url);
+    print(re.body);
+    print("long task done");
+  }
+
+  ///  *********************************************
+  ///     NOTIFICATION CREATION METHODS
+  ///  *********************************************
+  ///
+  static Future<void> createNewNotification() async {
+    bool isAllowed = await AwesomeNotifications().isNotificationAllowed();
+    if (!isAllowed) isAllowed = await displayNotificationRationale();
+    if (!isAllowed) return;
+
+    await AwesomeNotifications().createNotification(
+        content: NotificationContent(
+            id: -1, // -1 is replaced by a random number
+            channelKey: 'alerts',
+            title: 'Huston! The eagle has landed!',
+            body:
+                "A small step for a man, but a giant leap to Flutter's community!",
+            bigPicture: 'https://storage.googleapis.com/cms-storage-bucket/d406c736e7c4c57f5f61.png',
+            largeIcon: 'https://storage.googleapis.com/cms-storage-bucket/0dbfcc7a59cd1cf16282.png',
+            //'asset://assets/images/balloons-in-sky.jpg',
+            notificationLayout: NotificationLayout.BigPicture,
+            payload: {'notificationId': '1234567890'}),
+        actionButtons: [
+          NotificationActionButton(key: 'REDIRECT', label: 'Redirect'),
+          NotificationActionButton(
+              key: 'REPLY',
+              label: 'Reply Message',
+              requireInputText: true,
+              actionType: ActionType.SilentAction
+          ),
+          NotificationActionButton(
+              key: 'DISMISS',
+              label: 'Dismiss',
+              actionType: ActionType.DismissAction,
+              isDangerousOption: true)
+        ]);
+  }
+
+  static Future<void> scheduleNewNotification() async {
+    bool isAllowed = await AwesomeNotifications().isNotificationAllowed();
+    if (!isAllowed) isAllowed = await displayNotificationRationale();
+    if (!isAllowed) return;
+
+    await AwesomeNotifications().createNotification(
+        content: NotificationContent(
+            id: -1, // -1 is replaced by a random number
+            channelKey: 'alerts',
+            title: "Huston! The eagle has landed!",
+            body:
+                "A small step for a man, but a giant leap to Flutter's community!",
+            bigPicture: 'https://storage.googleapis.com/cms-storage-bucket/d406c736e7c4c57f5f61.png',
+            largeIcon: 'https://storage.googleapis.com/cms-storage-bucket/0dbfcc7a59cd1cf16282.png',
+            //'asset://assets/images/balloons-in-sky.jpg',
+            notificationLayout: NotificationLayout.BigPicture,
+            payload: {
+              'notificationId': '1234567890'
+            }),
+        actionButtons: [
+          NotificationActionButton(key: 'REDIRECT', label: 'Redirect'),
+          NotificationActionButton(
+              key: 'DISMISS',
+              label: 'Dismiss',
+              actionType: ActionType.DismissAction,
+              isDangerousOption: true)
+        ],
+        schedule: NotificationCalendar.fromDate(
+            date: DateTime.now().add(const Duration(seconds: 10))));
+  }
+
+  static Future<void> resetBadgeCounter() async {
+    await AwesomeNotifications().resetGlobalBadge();
+  }
+
+  static Future<void> cancelNotifications() async {
+    await AwesomeNotifications().cancelAll();
+  }
+}
+
+///  *********************************************
+///     MAIN WIDGET
+///  *********************************************
+///
+class MyApp extends StatefulWidget {
+  const MyApp({super.key});
 
   // The navigator key is necessary to navigate using static methods
   static final GlobalKey<NavigatorState> navigatorKey =
       GlobalKey<NavigatorState>();
 
+  static Color mainColor = const Color(0xFF9D50DD);
+
   @override
-  _AppState createState() => _AppState();
+  State<MyApp> createState() => _AppState();
 }
 
-class _AppState extends State<App> {
+class _AppState extends State<MyApp> {
+  // This widget is the root of your application.
+
+  static const String routeHome = '/', routeNotification = '/notification-page';
+
   @override
   void initState() {
+    NotificationController.startListeningNotificationEvents();
     super.initState();
+  }
 
-    // Only after at least the action method is set, the notification events are delivered
-    NotificationsController.initializeNotificationsEventListeners();
+  List<Route<dynamic>> onGenerateInitialRoutes(String initialRouteName) {
+    List<Route<dynamic>> pageStack = [];
+    pageStack.add(MaterialPageRoute(
+        builder: (_) =>
+            const MyHomePage(title: 'Awesome Notifications Example App')));
+    if (initialRouteName == routeNotification &&
+        NotificationController.initialAction != null) {
+      pageStack.add(MaterialPageRoute(
+          builder: (_) => NotificationPage(
+              receivedAction: NotificationController.initialAction!)));
+    }
+    return pageStack;
+  }
+
+  Route<dynamic>? onGenerateRoute(RouteSettings settings) {
+    switch (settings.name) {
+      case routeHome:
+        return MaterialPageRoute(
+            builder: (_) =>
+                const MyHomePage(title: 'Awesome Notifications Example App'));
+
+      case routeNotification:
+        ReceivedAction receivedAction = settings.arguments as ReceivedAction;
+        return MaterialPageRoute(
+            builder: (_) => NotificationPage(receivedAction: receivedAction));
+    }
+    return null;
   }
 
   @override
   Widget build(BuildContext context) {
-    String initialRoute =
-        App.initialCallAction == null ? PAGE_HOME : PAGE_PHONE_CALL;
-    debugPrint('initialRoute: $initialRoute');
     return MaterialApp(
-      title: App.name,
-      color: App.mainColor,
-      navigatorKey: App.navigatorKey,
-      initialRoute: initialRoute,
-      routes: materialRoutes,
-      builder: (context, child) => MediaQuery(
-        data: MediaQuery.of(context).copyWith(alwaysUse24HourFormat: true),
-        child: child ?? const SizedBox.shrink(),
+      title: 'Awesome Notifications - Simple Example',
+      navigatorKey: MyApp.navigatorKey,
+      onGenerateInitialRoutes: onGenerateInitialRoutes,
+      onGenerateRoute: onGenerateRoute,
+      theme: ThemeData(
+        primarySwatch: Colors.deepPurple,
       ),
-      theme: ThemesController.currentTheme,
+    );
+  }
+}
+
+///  *********************************************
+///     HOME PAGE
+///  *********************************************
+///
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key, required this.title});
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: const <Widget>[
+            Text(
+              'Push the buttons below to create new notifications',
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: Padding(
+        padding: const EdgeInsets.all(20.0),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            const SizedBox(width: 20),
+            FloatingActionButton(
+              heroTag: '1',
+              onPressed: () => NotificationController.createNewNotification(),
+              tooltip: 'Create New notification',
+              child: const Icon(Icons.outgoing_mail),
+            ),
+            const SizedBox(width: 10),
+            FloatingActionButton(
+              heroTag: '2',
+              onPressed: () => NotificationController.scheduleNewNotification(),
+              tooltip: 'Schedule New notification',
+              child: const Icon(Icons.access_time_outlined),
+            ),
+            const SizedBox(width: 10),
+            FloatingActionButton(
+              heroTag: '3',
+              onPressed: () => NotificationController.resetBadgeCounter(),
+              tooltip: 'Reset badge counter',
+              child: const Icon(Icons.exposure_zero),
+            ),
+            const SizedBox(width: 10),
+            FloatingActionButton(
+              heroTag: '4',
+              onPressed: () => NotificationController.cancelNotifications(),
+              tooltip: 'Cancel all notifications',
+              child: const Icon(Icons.delete_forever),
+            ),
+          ],
+        ),
+      ), // This trailing comma makes auto-formatting nicer for build methods.
+    );
+  }
+}
+
+///  *********************************************
+///     NOTIFICATION PAGE
+///  *********************************************
+///
+class NotificationPage extends StatelessWidget {
+  const NotificationPage({Key? key, required this.receivedAction})
+      : super(key: key);
+
+  final ReceivedAction receivedAction;
+
+  @override
+  Widget build(BuildContext context) {
+    bool hasLargeIcon = receivedAction.largeIconImage != null;
+    bool hasBigPicture = receivedAction.bigPictureImage != null;
+    double bigPictureSize = MediaQuery.of(context).size.height * .4;
+    double largeIconSize =
+        MediaQuery.of(context).size.height * (hasBigPicture ? .12 : .2);
+
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(receivedAction.title ?? receivedAction.body ?? ''),
+      ),
+      body: SingleChildScrollView(
+        padding: EdgeInsets.zero,
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.start,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            SizedBox(
+                height:
+                    hasBigPicture ? bigPictureSize + 40 : largeIconSize + 60,
+                child: hasBigPicture
+                    ? Stack(
+                        children: [
+                          if (hasBigPicture)
+                            FadeInImage(
+                              placeholder: const NetworkImage(
+                                  'https://cdn.syncfusion.com/content/images/common/placeholder.gif'),
+                              //AssetImage('assets/images/placeholder.gif'),
+                              height: bigPictureSize,
+                              width: MediaQuery.of(context).size.width,
+                              image: receivedAction.bigPictureImage!,
+                              fit: BoxFit.cover,
+                            ),
+                          if (hasLargeIcon)
+                            Positioned(
+                              bottom: 15,
+                              left: 20,
+                              child: ClipRRect(
+                                borderRadius: BorderRadius.all(
+                                    Radius.circular(largeIconSize)),
+                                child: FadeInImage(
+                                  placeholder: const NetworkImage(
+                                      'https://cdn.syncfusion.com/content/images/common/placeholder.gif'),
+                                  //AssetImage('assets/images/placeholder.gif'),
+                                  height: largeIconSize,
+                                  width: largeIconSize,
+                                  image: receivedAction.largeIconImage!,
+                                  fit: BoxFit.cover,
+                                ),
+                              ),
+                            )
+                        ],
+                      )
+                    : Center(
+                        child: ClipRRect(
+                          borderRadius:
+                              BorderRadius.all(Radius.circular(largeIconSize)),
+                          child: FadeInImage(
+                            placeholder: const NetworkImage(
+                                'https://cdn.syncfusion.com/content/images/common/placeholder.gif'),
+                            //AssetImage('assets/images/placeholder.gif'),
+                            height: largeIconSize,
+                            width: largeIconSize,
+                            image: receivedAction.largeIconImage!,
+                            fit: BoxFit.cover,
+                          ),
+                        ),
+                      )),
+            Padding(
+              padding: const EdgeInsets.only(bottom: 20.0, left: 20, right: 20),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.start,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  RichText(
+                      text: TextSpan(children: [
+                    if (receivedAction.title?.isNotEmpty ?? false)
+                      TextSpan(
+                        text: receivedAction.title!,
+                        style: Theme.of(context).textTheme.titleLarge,
+                      ),
+                    if ((receivedAction.title?.isNotEmpty ?? false) &&
+                        (receivedAction.body?.isNotEmpty ?? false))
+                      TextSpan(
+                        text: '\n\n',
+                        style: Theme.of(context).textTheme.bodyMedium,
+                      ),
+                    if (receivedAction.body?.isNotEmpty ?? false)
+                      TextSpan(
+                        text: receivedAction.body!,
+                        style: Theme.of(context).textTheme.bodyMedium,
+                      ),
+                  ]))
+                ],
+              ),
+            ),
+            Container(
+              color: Colors.black12,
+              padding: const EdgeInsets.all(20),
+              width: MediaQuery.of(context).size.width,
+              child: Text(receivedAction.toString()),
+            ),
+          ],
+        ),
+      ),
     );
   }
 }